Fixed log line integer warning

+++ IdentityCore/src/controllers/MSIDLocalInteractiveController.m
@@ -82,7 +82,7 @@
     {
         MSIDRequestCompletionBlock completionBlockWrapper = ^(MSIDTokenResult * _Nullable result, NSError * _Nullable error)
         {
-            MSID_LOG_INFO(self.requestParameters, @"Interactive flow finished result %@, error: %ld error domain: %@", _PII_NULLIFY(result), error.code, error.domain);
+            MSID_LOG_INFO(self.requestParameters, @"Interactive flow finished result %@, error: %ld error domain: %@", _PII_NULLIFY(result), (long)error.code, error.domain);
             completionBlock(result, error);
         };
